hw2/python_app/requirements.txt
rich
psycopg2-binary
pytest
